fix bugs

+++ src/attacks/get_networks.py
@@ -66,15 +66,14 @@ class get_networks:
         filepath = os.path.expanduser("~/output-01.csv")
         while not self._event_stop.is_set():
             try:
-                if not globals.send_beacon:
+                if not globals.send_beacon and not globals.stop_scan:
                     if globals.proc:
                         globals.proc.terminate()
                         globals.proc.wait()
-                    if globals.retry_time_left == 0:
-                        if not globals.stop_scan:
-                            globals.proc = self.run_airodump(globals.interface, globals.channel)
-                        elif globals.attack_menu and not globals.guided_deauth:
+                        if globals.attack_menu and not globals.guided_deauth:
                             globals.proc = self.run_airodump(globals.interface, channel=globals.channel, bssid=globals.selected_bssid)
+                        else:
+                            globals.proc = self.run_airodump(globals.interface, globals.channel)
                         if globals.proc:
                             globals.retry_time_left = 10
                             for _ in range(10):
@@ -97,11 +96,11 @@ class get_networks:
                                 os.remove(filepath)
                             if globals.fix:
                                 globals.csv_saver.log()
-                        else:
-                            if globals.proc:
-                                globals.proc.terminate()
-                                globals.proc.wait()
-                            time.sleep(1)
+                else:
+                    if globals.proc:
+                        globals.proc.terminate()
+                        globals.proc.wait()
+                    time.sleep(1)
             except Exception:
                 globals.proc.terminate()
                 time.sleep(1)
